requirements.txt
Django==5.1.6
djangorestframework==3.15.2
psycopg2==2.9.10
djangorestframework_simplejwt==5.4.0
pytest>=7.0.0
pytest-django>=4.0.0